Refactor equipable Items + some equipment features (#5)
* Remove stats columns of Character

Stength, Dexterity, Intelligence and Vitality will be calculated with the class, race, level and equipment of the character. Still retreivable by Gql but not stored in database anymore.

* Add Shoulder armor position

* WIP: refactor equipables entities

* WIP: refactor equipment handling

Equipment now have a list of items.
ParentEntity Item now have equipmentPosition and equipable property.

* Feat: equip equipment pieces

Character can now equip weapons and armors. These must be in the character inventory, and if a piece is already equipped at the current position, it go back to inventory.